requirements.txt
# AutoStream AI Agent - Dependencies
# Python 3.9+

# ── Core LangChain / LangGraph ────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0

# ── Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── Web / API ──────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn>=0.23.0
